ON-4985 # Changed region to tenant
