+++ classes/TeamMembers.js
@@ -2,12 +2,12 @@
 'use strict'
 
 const OneBlinkAPI = require('../lib/one-blink-api.js')
-const getRegionUrl = require('../lib/region')
+const getTenantUrl = require('../lib/tenant')
 module.exports = class TeamMembers extends OneBlinkAPI {
   constructor(options /* : ConstructorOptions */) {
     options = options || {}
     super(
-      getRegionUrl(options.regionCode, options.oneBlinkAPIOrigin),
+      getTenantUrl(options.tenant, options.oneBlinkAPIOrigin),
       options.accessKey,
       options.secretKey
     )
